samples: updates v1beta3 to v1; renames v1beta2 samples (#198)
