+++ document-ai/test/parse-table.v1beta2.test.js
@@ -24,18 +24,15 @@ const execSync = cmd => cp.execSync(cmd, {encoding: 'utf-8'});
 const cwd = path.join(__dirname, '..');
 const projectId = process.env.GCLOUD_PROJECT;
 const LOCATION = 'us';
-const MODEL_NAME =
-  process.env.MODEL_NAME ||
-  'projects/1046198160504/locations/us-central1/models/TCN7483069430457434112';
 
-describe('Document AI parse with AutoML model', () => {
-  it('should run use an AutoML model to parse a PDF', async () => {
+describe('Document AI parse table', () => {
+  it('should parse the GCS invoice example as as table (v1beta2)', async () => {
     const stdout = execSync(
-      `node ./parse_with_model.js ${projectId} ${LOCATION} ${MODEL_NAME}`,
+      `node ./parse-table.v1beta2.js ${projectId} ${LOCATION}`,
       {
         cwd,
       }
     );
-    assert.match(stdout, /Label/);
+    assert.match(stdout, /Header row/);
   });
 });
