--- document-ai/test/set_endpoint.test.js
@@ -1,35 +1,0 @@
-/**
- * Copyright 2019, Google, Inc.
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-'use strict';
-
-const path = require('path');
-const {assert} = require('chai');
-const cp = require('child_process');
-
-const execSync = cmd => cp.execSync(cmd, {encoding: 'utf-8'});
-
-const cwd = path.join(__dirname, '..');
-const projectId = process.env.GCLOUD_PROJECT;
-const LOCATION = 'eu';
-
-describe('Document AI set endpoint', () => {
-  it('should process a PDF in another region', async () => {
-    const stdout = execSync(`node ./set_endpoint.js ${projectId} ${LOCATION}`, {
-      cwd,
-    });
-    assert.match(stdout, /Entity/);
-  });
-});
